==> scripts/03_monthly_aggreagation.ipynb <==
import pandas as pd
import os

enrolment_agg_df=pd.read_csv("../final_dataset/enrolment_aggregated.csv",parse_dates=["date"])

demographic_agg_df=pd.read_csv("../final_dataset/demographic_aggregated.csv",parse_dates=["date"])

enrolment_agg_df["year_month"]=(enrolment_agg_df["date"].dt.to_period("M").astype(str))

demographic_agg_df["year_month"]=(
  demographic_agg_df["date"].dt.to_period("M").astype(str)
)

enrolment_monthly_df=(
  enrolment_agg_df.groupby(["year_month","state_final","district_norm"],as_index=False).agg({"age_0_5":"sum","age_5_17":"sum","age_18_greater":"sum","total_enrolments":"sum"})
)

demographic_monthly_df=(demographic_agg_df.groupby(["year_month","state_final","district_norm"],as_index=False).agg({"demo_age_5_17":"sum","demo_age_17_":"sum","total_demographic_updates":"sum"}))

print("Original enrolments:",enrolment_agg_df["total_enrolments"].sum())

print("Monthly enrolments:",enrolment_monthly_df["total_enrolments"].sum())

(enrolment_monthly_df["total_enrolments"]>=0).all

print("Original demographic updates:",demographic_agg_df["total_demographic_updates"].sum())

print("Monthly demographic updates:",demographic_monthly_df["total_demographic_updates"].sum())

(demographic_monthly_df["total_demographic_updates"]>=0).all

enrolment_monthly_df.head()

demographic_monthly_df.head()

os.makedirs("../final_dataset",exist_ok=True)
enrolment_monthly_df.to_csv("../final_dataset/enrolment_monthly.csv",index=False)
demographic_monthly_df.to_csv("../final_dataset/demographic_monthly.csv",index=False)

os.listdir("../final_dataset")
